Fixed an issue where the cell.marked would only be set for cells within the current month.
This allows for subclasses to check whether the cell is marked and
return different value for customizing the appearance of the cell based
on whether its marked.
Also added re-setting the cell's background color every time
collectionView:(UICollectionView *)collectionView
cellForItemAtIndexPath:(NSIndexPath *)indexPath is called, so that
subclasses can return a different value for 'selfBackgroundColor' and
get that color set on the cell.